+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 0.12.0-dev
 - BC break: renamed IXmlConvertible to XmlConvertible and IRssExtension to RssExtension
+- BC break: replaced Nette events with PSR-14 event dispatcher
 
 Version 0.11.0
 - made Generator::$dataSource readable